pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aerich==0.8.2",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.15",
    "aioredis==2.0.1",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "async-timeout==5.0.1",
    "asyncclick==8.1.8",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "babel==2.17.0",
    "bcrypt==4.0.1",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "dictdiffer==0.9.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "environs==14.1.1",
    "fastadmin[fastapi,tortoise-orm]==0.2.19",
    "fastapi==0.115.12",
    "frozenlist==1.5.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "iso8601==2.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "packaging==24.2",
    "passlib[bcrypt]==1.7.4",
    "pendulum==3.0.0",
    "propcache==0.3.1",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.1",
    "pydantic-core==2.33.0",
    "pyjwt==2.10.1",
    "pypika-tortoise==0.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis==5.2.1",
    "requests==2.32.3",
    "ruff>=0.12.2",
    "setuptools==78.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.1",
    "time-machine==2.16.0",
    "tomli-w==1.2.0",
    "tomlkit==0.13.2",
    "tortoise-orm==0.24.2",
    "typing-extensions==4.13.0",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
    "yarl==1.18.3",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.28.1",
    "tortoise-orm[asyncpg]>=0.24.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.aerich]
tortoise_orm = "app.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "pytest-anyio>=0.0.0",
]
# Fix for aerich 0.8.2 bug with old_models being None
# This ensures proper model loading
